forloop
Gave the array values a data-value.  When a user clicks on an answer, it
will fetch the value of the array object, put it through a for loop, and
return the results through the console.

If the answer is array [3] and the user clicks the 4th answer, it will
only print out one answer.  Now I need to make a function that sees how
many results it turns out.  If it only turns out one result, the user
got the question right.